+++ lib/Dialect/TTL/Transforms/BufferizableOpInterfaceImpl.cpp
@@ -8,6 +8,7 @@
 #include "ttlang/Dialect/TTL/IR/TTLOps.h"
 
 #include "mlir/Dialect/Bufferization/IR/BufferizableOpInterface.h"
+#include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/DialectRegistry.h"
 #include "mlir/IR/MLIRContext.h"
 
@@ -21,6 +22,7 @@ using bufferization::AnalysisState;
 using bufferization::BufferizableOpInterface;
 using bufferization::BufferizationOptions;
 using bufferization::BufferizationState;
+using bufferization::BufferLikeType;
 using bufferization::BufferRelation;
 
 /// ttl.attach_cb is a pure aliasing op: it forwards its tensor/memref operand
@@ -59,6 +61,18 @@ struct AttachCBOpInterface
     return true;
   }
 
+  FailureOr<BufferLikeType>
+  getBufferType(Operation *op, Value value, const BufferizationOptions &options,
+                const BufferizationState &state,
+                SmallVector<Value> &invocationStack) const {
+    auto attachOp = cast<AttachCBOp>(op);
+    if (value == attachOp.getResult()) {
+      return bufferization::getBufferType(attachOp.getTensor(), options, state,
+                                          invocationStack);
+    }
+    return bufferization::getBufferType(value, options, state, invocationStack);
+  }
+
   LogicalResult bufferize(Operation *op, RewriterBase &rewriter,
                           const BufferizationOptions &options,
                           BufferizationState &state) const {
@@ -74,12 +88,155 @@ struct AttachCBOpInterface
   }
 };
 
+template <typename OpTy>
+struct CBViewOpInterface
+    : public BufferizableOpInterface::ExternalModel<CBViewOpInterface<OpTy>,
+                                                    OpTy> {
+  bool bufferizesToMemoryRead(Operation *, OpOperand &,
+                              const AnalysisState &) const {
+    return false;
+  }
+
+  bool bufferizesToMemoryWrite(Operation *, OpOperand &,
+                               const AnalysisState &) const {
+    return false;
+  }
+
+  AliasingValueList getAliasingValues(Operation *, OpOperand &,
+                                      const AnalysisState &) const {
+    return {};
+  }
+
+  BufferRelation bufferRelation(Operation *, OpOperand &,
+                                const AnalysisState &) const {
+    return BufferRelation::Unknown;
+  }
+
+  bool isWritable(Operation *, Value, const AnalysisState &) const {
+    return true;
+  }
+
+  LogicalResult bufferize(Operation *op, RewriterBase &rewriter,
+                          const BufferizationOptions &options,
+                          BufferizationState &) const {
+    auto cbOp = cast<OpTy>(op);
+    Type resultType = cbOp->getResult(0).getType();
+    if (mlir::isa<BaseMemRefType>(resultType)) {
+      return mlir::success();
+    }
+
+    auto tensorType = mlir::dyn_cast<RankedTensorType>(resultType);
+    if (!tensorType) {
+      return op->emitError()
+             << "expected ranked tensor result prior to bufferization";
+    }
+
+    FailureOr<BaseMemRefType> bufferType =
+        bufferization::getMemRefType(tensorType, options);
+    if (failed(bufferType)) {
+      return failure();
+    }
+
+    rewriter.replaceOpWithNewOp<OpTy>(cbOp, *bufferType, cbOp.getCb());
+    return success();
+  }
+};
+
+struct CopyOpInterface
+    : public BufferizableOpInterface::ExternalModel<CopyOpInterface, CopyOp> {
+  bool bufferizesToMemoryRead(Operation *op, OpOperand &opOperand,
+                              const AnalysisState &) const {
+    auto copy = cast<CopyOp>(op);
+    if (opOperand.getOperandNumber() != 0) {
+      return false;
+    }
+    return !copy.isSrcCircularBuffer();
+  }
+
+  bool bufferizesToMemoryWrite(Operation *op, OpOperand &opOperand,
+                               const AnalysisState &) const {
+    auto copy = cast<CopyOp>(op);
+    if (opOperand.getOperandNumber() != 1) {
+      return false;
+    }
+    return copy.isSrcCircularBuffer();
+  }
+
+  AliasingValueList getAliasingValues(Operation *, OpOperand &,
+                                      const AnalysisState &) const {
+    return {};
+  }
+
+  BufferRelation bufferRelation(Operation *, OpOperand &,
+                                const AnalysisState &) const {
+    return BufferRelation::Unknown;
+  }
+
+  bool isWritable(Operation *, Value, const AnalysisState &) const {
+    return true;
+  }
+
+  LogicalResult bufferize(Operation *op, RewriterBase &rewriter,
+                          const BufferizationOptions &options,
+                          BufferizationState &state) const {
+    auto copy = cast<CopyOp>(op);
+    Value newSrc = copy.getSrc();
+    Value newDst = copy.getDst();
+
+    auto convertIfTensor = [&](Value oldVal,
+                               Value &newVal) -> FailureOr<RankedTensorType> {
+      auto tensorTy = mlir::dyn_cast<RankedTensorType>(oldVal.getType());
+      if (!tensorTy) {
+        return RankedTensorType();
+      }
+      FailureOr<Value> buffer =
+          bufferization::getBuffer(rewriter, oldVal, options, state);
+      if (failed(buffer)) {
+        return failure();
+      }
+      newVal = *buffer;
+      return tensorTy;
+    };
+
+    FailureOr<RankedTensorType> newTensorTy;
+    if (!copy.isSrcCircularBuffer()) {
+      newTensorTy = convertIfTensor(copy.getSrc(), newSrc);
+      if (failed(newTensorTy)) {
+        return failure();
+      }
+    }
+    if (!copy.isDstCircularBuffer()) {
+      auto converted = convertIfTensor(copy.getDst(), newDst);
+      if (failed(converted)) {
+        return failure();
+      }
+      if (*converted) {
+        newTensorTy = converted;
+      }
+    }
+
+    TypeAttr tensorAttr;
+    if (*newTensorTy) {
+      tensorAttr = TypeAttr::get(*newTensorTy);
+    } else {
+      tensorAttr = copy.getTensorTypeAttr();
+    }
+    rewriter.replaceOpWithNewOp<CopyOp>(copy, copy.getResult().getType(),
+                                        newSrc, newDst, tensorAttr);
+
+    return success();
+  }
+};
+
 } // namespace
 
 void registerBufferizableOpInterfaceExternalModels(DialectRegistry &registry) {
   registry.addExtension<TTLDialect>(+[](MLIRContext *ctx, TTLDialect *dialect) {
     (void)dialect;
     AttachCBOp::attachInterface<AttachCBOpInterface>(*ctx);
+    CBReserveOp::attachInterface<CBViewOpInterface<CBReserveOp>>(*ctx);
+    CBWaitOp::attachInterface<CBViewOpInterface<CBWaitOp>>(*ctx);
+    CopyOp::attachInterface<CopyOpInterface>(*ctx);
   });
 }